Fix usage of the crypto_sign_open() API
The crypto_sign_open() function accepts the signature and message as a
single parameter. The first parameter is a scratch buffer it uses to
compute the result. The original implementation was trying to pass the
message in place of the scratch buffer, which is wrong.

This change corrects the previous oversight. Running under valgrind now
detects no memory corruption.